refactor: split lib into several command packages
